Add option to link against libc++ and test compile that as well
Using libc++ has the adavantage that a lot of old C++03 comaptibility
features are not included, i.e. std::binary_function etc.

Compiling against that does therefore allow us to keep the codebase in
better shape.

By using add_link_options we can avoid having to touch every target.
Although this feature is cmake 3.13 and newer only we deem that okay as
it is a CI feature only.