log4tango/LoggerStream: Define copy and move operators explicitly
This is more explicit and let us avoid deprecation warnings.

We delete all of them except the move constructor. As that is special.

The move constructor is used in the Logger class like

  inline LoggerStream debug_stream (void) {
    return LoggerStream(*this, Level::DEBUG, true);
  }

so without copy elision this requires LoggerStream to be either
move-constructable or copy-constructable.

We can't just use the default version of the move constructor as that
would leave the pointer member _buffer dangling. The reason this is
currently no problem is that the move is optimized away due to RVO.

So we spell out the move constructor *and* set _buffer to nulltpr.