cppapi/server/idl: Make code generation from IDL file a target
This avoids executing it everytime on a build.

The code to patch the IDL file is moved to a separate cmake file so that
it can be executed as part of the target. And the targets which rely on
this now just need to be dependent on idl_ojects, if they need it
compiled, or just idl_source if they want to ensure that the C++ code
is present.