configure/CMakeLists.txt: Enforce a minimum libzmq version
The autotools build system of the tango source distribution already
enforces a minimum libzmq version of 4.0.5.

We now do the same for the CMake build system.

This is a pure compile time check, so it works with cross compiling. And
as it does not use pkg-config, it can also later be ported to other
non-linux platforms like Windows.

Close #788.