configure/CMakeLists.txt: Select correct MSVC runtime on Windows
We currently feed in the /MT or /MTd flags for compiling the static
library on Windows when compiling in appveyor.

But this makes it hard to compile locally.

So let's derive the correct MSVC runtime from BUILD_SHARED_LIBS.

Strictly speaking MSVC can be linked statically when we build a shared
library and vice versa, but we don't support that.

See also https://docs.microsoft.com/en-us/cpp/c-runtime-library/crt-library-features?view=msvc-160.