configure/cmake_win.cmake: Install tango_const.h as well
This header file is the result of configuring tango_const.h.in.

Broken since forever in the cmake build on windows when packaging a
cpack zip file.

For *nix OSes and plain installations the cmake install rule in
cppapi/server/CMakeLists.txt installs that file as well.

Due to a cmake/cpack "feature" where absolute paths are not supported for
DESTINATION in install commands, these duplicated install commands are
necessary.