.gitlab-ci.yml: Add job for testing custom locations and minimum versions for dependent libraries
cppTango supports a variety of XXX_BASE cmake options to define custom
locations for omniORB, libzmq, cppzmq and tango-idl.

But we currently don't have a test which verifies that this works.

Let's change that.

And while we are at it, we also use the lowest possible version of these
libraries to verify that it at least compiles.