cppapi/server/dserversignal.h: Support musl libc
Alpine linux uses musl as libc and in that case __GLIBC__ is not defined, but
nothing else matches either.

Close code inspection reveals that we always define the function on all
platforms. So let's just do that.

The fix in main, 918e6459 (Remove auth_signal() function... and
associated dead code Following a suggestion from @mliszcz, 2020-04-28),
removed that function, but we want to keep the ABI stable here.