log4tango/config/config.cmake: Fix threading support detection
On Apline Linux the CMAKE_THREAD_LIBS_INIT variable is empty as thread
support is built into musl.

This results in no threading support detected. We can workaround this by
treating HAVE_PTHREAD_CREATE as enough for threading support as well.

It would be better to use the Threads_Found from FindThreads but that is
only available starting with cmake 3.17 [1].

[1]: https://cmake.org/cmake/help/v3.17/module/FindThreads.html