.gitlab-ci.yml: Fix cppzmq version in some jobs
And in df11a078 (.gitlab-ci.yml: Raise cppzmq versions and use non-stock where needed, 2021-06-25) we raised the required cppzmq version.

But as we did the MR merging yesterday in a binge session this MRs didn't
pay attention to other MRs.

So let's use the correct cppzmq version also for other jobs as well.