devapi_data.cpp/operator<<: Avoid using TANGO_LONG32
On Windows with MSVC [1] long long is 8 bytes large and as TANGO_LONG32
is always defined on *nix, where it is also 8 bytes large,  we have
together with the current definition of TANGO_LONG32 the case that we
always use a 8 byte type.

So let's just use Tango::DevLong64 instead.

[1]: https://docs.microsoft.com/en-us/cpp/cpp/data-type-ranges?view=msvc-160