cppapi/client/filedatabase.cpp/DbInfo: Avoid using TANGO_LONG64
The use of TANGO_LONG64 is wrong here as %d is a flag for a signed
integer and not an unsigned one has used here.

By rewriting the code usig C++11 style std::to_string we avoid having to
deal with these issue at all.

And using lambda function for repeated purposes makes the code also less
repetitive.