Add unit tests for jpeg encoding.
Test is configured via cmake to access the resources, hence the
modification in cxxtest/CMakelists.txt to find configured file, that is
in the build tree, and not the source one.

Modification of the run with fixtures script to allow not to use the
fixture (allow for this test to run much faster).

Image used is peppers from
http://sipi.usc.edu/database/database.php?volume=misc. The image is free
of right. Gimp was used to extract raw data from the input tiff file.