Fix crash in attr read with no value on IDLv1/v2
If we detected an attribute with a valid quality but no value set while
storing attribute sequences for network transfer, we released previously
allocated sequences before throwing DevFailed.

This release was not needed because previously allocated sequences were
already released in earlier iterations of this loop, immediately after
they were stored in the `back` list (AttributeValueList) some 60 lines
below. This resulted in a double-delete and a crash of the device server.

This issue impacts only IDLv1/v2 devices.

Fixes #636.