DeviceAttribute: Allow extracting the DEV_STATE attribute multiple times
Since forever in the git history we treat the state attribute of type
DEVICE_STATE special.

The DEVICE_STATE attribute is stored in AttrValUnion::dev_state_attr
whereas other attributes of type ATT_STATE are stored in
AttrValUnion::state_att_value.

When we convert in ApiUtil::attr_to_device/attr_to_device_base between
the CORBA types AttributeValue_XXX and Tango::DeviceAttribute we also
make a distinction and store DEVICE_STATE values in
DeviceAttribute::d_state and other ATT_STATE attributes in
DeviceAttribute::StateSeq.

It is unclear at this point why this distinction was made.

Now the problem is that we store what type of state entry is present in
DeviceAttribute::d_state_filled, which is true for DEVICE_STATE and false
for all other types, *and* clear that flag again when extracting
attributes of type DEV_STATE.

The result of clearing d_state_filled is that you can not extract a
DEV_STATE attribute twice.

By removing the clearing of d_state_filled we now allow to extracting
the same DEV_STATE attribute multiple times.