Add a global flag indicating library threads
We need to differentiate between library and external threads for the
purpose of logging subdevices (which can only be set from the library
threads).

Currently presence of PyData structure indicates a library thread but
this structure is no longer used and will be removed. We also want to
reduce dependency on omnithread features, and a thread_local variable
is used instead.