CMake: Fix and unify install commands
On Windows we currently install the tango headers twice, once in include
and once in include/tango. But some files are only in one of them as well.

The reasons are twofold. The first is that install statements for
windows-only files are missing. These were added now.

The second is that install statements using an absolute destination path
are ignored when using cpack [1]. This results in differences when
creating a ZIP package versus installing locally.

The incomplete solution to that problem was, that we
always had duplicated install statements, one for *nix with an absolute
path via CMAKE_INSTALL_FULL_XXX and one for windows/cpack with a
relative path.

A better solution is to use CMAKE_INSTALL_XXX which is a relative path,
and thus works on *nix *and* on windows/cpack.

We now always install all header files in include/tango. We have always
done that on *nix but for windows users this now requires to also add
include/tango as search directory for headers.

[1]: https://cmake.org/cmake/help/v3.23/command/install.html