Let other choose the C++ standard
In Linux ecosystem the distribution developers choose a sane value for
the C++ standard offered by default from compilers available in the distribution.
That choice can be affected by a lot of factors (which standard used to build
the other C++ code in the distribution, bugs or missing support in glibc or
in the STL and so on). We can't know what is the best C++ standard to use to build
our code so the best thing that we can do here is let this choice
to distribution developers.