Check the requested C++ standard and not one of those available
There is no point to check compiler version available. We would want
provide the right flavour of the code according C++ standard requested
by user and not one of those available on the system. In other words
it must be possible build the C++98 flavour on a C++11 capable system
thing that isn't possible at the moment. So check C++ standard requested
and not the compiler version.