Windows: Add missing dllimport statements
When building a shared library on windows using Visual Studio one has to
declare what should be exported from the DLL.

We already set CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS=ON in cmake, but that actually
still requires global data symbols to be marked with `__declspec(dllimport)` [1].

Let's do that then.

The solution for omniORB does not require patching the generated code as we can just set a
prepocessor define to achieve that [2]. This is also the documented
solution.

Thanks to @liujin001 for the initial patch.

[1]: https://cmake.org/cmake/help/latest/prop_tgt/WINDOWS_EXPORT_ALL_SYMBOLS.html
[2]: https://omniorb.sourceforge.io/omni40/omniORB/omniORB012.html