appveyor.yml: Fix MSVC runtime library selection
Since 4b2063c3 (Rework static/shared variants into a configure-time
build-switch (#688), 2020-05-28) the default for the runtime library was
always static.

This has not proven to work in all combinations of shared library/static
library vs static MSVC/shared MSVC.

Although it is possible to mix shared library with static MSVC and static
library with shared MSVC, these combinations could not be made working
on this developers machine with comiling and running TangoTest. This is
attributed to the brittle compilation setup of the dependencies.

We therefore now default to the common options:
- Use static MSVC with static library
- And use dynamic MSVC with dynamic library

We don't hardcode the MSVC static/dynamic runtime flags in the CMake files
as that is something a developer should be able to override.