Tests/cpp_test_ds/main.cpp: Fix unclean server shutdown
In 8478cb3c (cpp_test_suite/old_tests: Remove cout/coutv macro handling,
2022-05-10) a fix was added which refetches the Tango::Util object if it
was not yet fetched.

But that does not work as Tango::Util::instance returns the current
instance which can also be null. In addition it also prints an error
message in this case.

As we only can do the server cleanup if we have an Util object, the fix is
to not do the cleanup in all other cases.