TANGO_THROW_ON_DEFAULT: Introduce and use it
A lot of switch statements currently have an empty default statement. In
most cases when we are, for example iterating over an enumeration, we know
that anything else than the listed cases are a programming error. So let's
use the macro TANGO_THROW_ON_DEFAULT in these cases.

To some switch statements we also added a "do nothing" block with
encountered values in the tests.