log4tango::LogInitiator: Fix static data member DLL export Fix #953
Replace the _end_log and _begin_log construct with static data members
with more of a tag construct.
It integrates better with windows where static data member must be
exported with __declspec on top of using the EXPORT_ALL_SYMBOLS cmake
variable on compilation.
the new _end_log and _begin_log could be inlined but c++17 support is
required, it shouldn't cause much trouble if they're not though.