CMakeLists.txt: Remove TANGO_USE_USING_NAMESPACE
That option was introdcued in 63089294 (Add CMAKE option to allow to not have using namespace in header files, 2019-02-04) to ease the transition between 9.3 where

using namespace std;

was done unconditionally in the tango.h header file and 9.4 where this is
not done anymore.

It was now realized that having an option which is deprecated by default
does not make a lot of sense.

So let's just remove it.

Users who are used to the old way can just do

 #include <tango.h>
 using namespace std;

and get the same behaviour as in 9.3.